+++ scripts/dzVents/runtime/integration-tests/stage2.lua
@@ -57,10 +57,9 @@ local testDimmer = function(name)
 	res = res and checkAttributes(dev, {
 		["id"] = 39,
 		["state"] = "On",
-		["lastLevel"] = 75;
+		["lastLevel"] = 75, -- this script is NOT triggered by the dimmer so lastLevel is the current level
 		["level"] = 75;
 	})
-	print(2222, otherdevices_lastlevel['vdSwitchDimmer'])
 	tstMsg('Test dimmer', res)
 	return res
 end
@@ -601,7 +600,7 @@ return {
 		res = res and testWaterflow("vdWaterflow")
 		res = res and testWind('vdWind')
 		res = res and testWind('vdWindTempChill')
---		res = res and testGroup('gpGroup')
+		res = res and testGroup('gpGroup')
 		res = res and testVariableInt('varInteger')
 		res = res and testVariableFloat('varFloat')
 		res = res and testVariableString('varString')