+++ scripts/dzVents/runtime/tests/tstData.lua
@@ -23,9 +23,10 @@ local testData = {
 				["hardwareType"] = "ht1",
 				["hardwareTypeValue"] = 1;
 				["hardwareID"] = 1,
+				['unit'] = 1,
 				heatingMode = '3'
 			};
-			["deviceID"] = "",
+			["deviceID"] = "__1",
 			["rawData"] = {
 				[1] = "1",
 				[2] = '2',
@@ -53,12 +54,13 @@ local testData = {
 				["hardwareType"] = "ht2",
 				["hardwareTypeValue"] = 2;
 				["hardwareID"] = 2,
+				['unit'] = 1,
 				temperature = 12,
 				dewpoint = 66,
 				humidity = 67,
 				--lux = 4
 			};
-			["deviceID"] = "",
+			["deviceID"] = "__2",
 			["rawData"] = {
 				[1] = "4",
 				[2] = '5',
@@ -86,10 +88,11 @@ local testData = {
 				["hardwareType"] = "ht3",
 				["hardwareTypeValue"] = 3;
 				["hardwareID"] = 3,
+				['unit'] = 1,
 				WActual = 11,
 				uv = 23
 			};
-			["deviceID"] = "",
+			["deviceID"] = "__3",
 			["rawData"] = {
 				[1] = "7",
 				[2] = '8',
@@ -121,13 +124,14 @@ local testData = {
 				["hardwareType"] = "ht4",
 				["hardwareTypeValue"] = 4;
 				["hardwareID"] = 4,
+				['unit'] = 1,
 				utility = 123,
 				weather = 'Nice',
 				rainLastHour = 12,
 				rain = 666,
 				setPoint = 10
 			};
-			["deviceID"] = "",
+			["deviceID"] = "__4",
 			["rawData"] = {
 				[1] = "10",
 				[2] = '11',
@@ -157,13 +161,14 @@ local testData = {
 				["hardwareType"] = "ht5",
 				["hardwareTypeValue"] = 5;
 				["hardwareID"] = 0,
+				['unit'] = 1,
 				WhTotal = 14,
 				WActual = 13,
 				WhToday = 1.234,
 				level = 10,
 				counterTotal = '567 kWh'
 			};
-			["deviceID"] = "",
+			["deviceID"] = "__5",
 			["rawData"] = {
 				[1] = "13",
 				[2] = '14',
@@ -190,9 +195,10 @@ local testData = {
 				["hardwareType"] = "ht4",
 				["hardwareTypeValue"] = 4;
 				["hardwareID"] = 4,
+				['unit'] = 1,
 				["value"] = 16.5, -- ?
 			};
-			["deviceID"] = "",
+			["deviceID"] = "__6",
 			["rawData"] = {},
 			["baseType"] = "device";
 			["changed"] = true;
@@ -216,9 +222,10 @@ local testData = {
 				["hardwareTypeValue"] = 0;
 				["hardwareID"] = 0,
 				["_state"] = 16.5,
+				['unit'] = 1,
 				WActual = 16.5
 			};
-			["deviceID"] = "",
+			["deviceID"] = "__7",
 			["rawData"] = {
 				[1] = "16.5"
 			},
@@ -244,9 +251,10 @@ local testData = {
 				["hardwareTypeValue"] = 0;
 				["hardwareID"] = 0,
 				["_state"] = 16.5,
+				['unit'] = 1,
 				text = 'Blah'
 			};
-			["deviceID"] = "",
+			["deviceID"] = "__8",
 			["rawData"] = {
 				[1] = "16.5"
 			},
@@ -273,9 +281,10 @@ local testData = {
 				["hardwareTypeValue"] = 0;
 				["hardwareID"] = 0,
 				["_state"] = 16.5,
+				['unit'] = 1,
 				text = 'Blah'
 			};
-			["deviceID"] = "",
+			["deviceID"] = "__9",
 			["rawData"] = {
 				[1] = "16.5"
 			},
@@ -302,9 +311,10 @@ local testData = {
 				["hardwareTypeValue"] = 0;
 				["hardwareID"] = 0,
 				["_state"] = 16.5,
+				['unit'] = 1,
 				text = 'Blah'
 			};
-			["deviceID"] = "",
+			["deviceID"] = "__10",
 			["rawData"] = {
 				[1] = "16.5"
 			},